OPENCAST-1250 LS-119 Limit small panning movements
This implementation if specific to the pan-only camera operator, and uses the alpha x value to
approximate the frame width (i.e. there isn't a separate configuration - the behaviour can be
changed by increasing or decreasing alpha x).

A fuller implementation of this needs a way to map the camera's current zoom to the frame width,
but as the pan-only operator uses a fixed zoom, this isn't required immediately.